refactor: simplify platform support in loop package
